# prowl/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Prowl LLM Orchestration API",
    "description": "REST API for LLM-driven orchestration of the Prowl security crawler. Enables programmatic crawl control, discovery queries, mid-crawl adjustments, intervention handling, and attack surface reporting.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/report/hindsight": {
      "get": {
        "tags": [
          "v1",
          "reports"
        ],
        "summary": "Hindsight analysis insights",
        "description": "Returns hindsight feedback insights: patterns detected from non-success responses (auth boundaries, method hints, etc.).",
        "operationId": "get_hindsight_api_v1_report_hindsight_get",
        "parameters": [
          {
            "name": "insight_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by insight type (auth_redirect, method_not_allowed, etc.)",
              "title": "Insight Type"
            },
            "description": "Filter by insight type (auth_redirect, method_not_allowed, etc.)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Hindsight Api V1 Report Hindsight Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}